+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased] - ReleaseDate
 
+### Features
+
+- [**BREAKING**] Support for SessionConfig::V2 has been moved behind the
+  `experimental-session-config` feature flag.
+  ([#300](https://github.com/matrix-org/vodozemac/pull/300)).
+
 ### Bug Fixes
 
 - [**BREAKING**] The `Account::create_inbound_session()` method now requires a